Foundations: Control Flow, Data Types #time 270m

+++ control-flow/solution/ifelse_integers.py
@@ -1,6 +1,6 @@
 number = input('What is your number?: ')
 
-if float(number) % 1:
+if int(number) % 1:
     print(False)
 else:
     print(True)
